+++ layers/CMSSchema/packages/custompostmeta/src/FieldResolvers/ObjectType/CustomPostObjectTypeFieldResolver.php
@@ -48,6 +48,11 @@ class CustomPostObjectTypeFieldResolver extends AbstractWithMetaObjectTypeFieldR
         return $this->getCustomPostMetaTypeAPI();
     }
 
+    protected function getLogicalCustomPostMetaTypeAPI(): CustomPostMetaTypeAPIInterface
+    {
+        return $this->getCustomPostMetaTypeAPI();
+    }
+
     /**
      * @return string[]
      */
@@ -72,7 +77,7 @@ class CustomPostObjectTypeFieldResolver extends AbstractWithMetaObjectTypeFieldR
         switch ($fieldDataAccessor->getFieldName()) {
             case 'metaKeys':
                 $metaKeys = [];
-                $customPostMetaTypeAPI = $this->getCustomPostMetaTypeAPI();
+                $customPostMetaTypeAPI = $this->getLogicalCustomPostMetaTypeAPI();
                 $allCustomPostMetaKeys = $customPostMetaTypeAPI->getCustomPostMetaKeys($customPost);
                 foreach ($allCustomPostMetaKeys as $key) {
                     if (!$customPostMetaTypeAPI->validateIsMetaKeyAllowed($key)) {
@@ -88,7 +93,7 @@ class CustomPostObjectTypeFieldResolver extends AbstractWithMetaObjectTypeFieldR
                     $objectTypeFieldResolutionFeedbackStore,
                 );
             case 'metaValue':
-                $metaValue = $this->getCustomPostMetaTypeAPI()->getCustomPostMeta(
+                $metaValue = $this->getLogicalCustomPostMetaTypeAPI()->getCustomPostMeta(
                     $customPost,
                     $fieldDataAccessor->getValue('key'),
                     true
@@ -99,7 +104,7 @@ class CustomPostObjectTypeFieldResolver extends AbstractWithMetaObjectTypeFieldR
                 }
                 return $metaValue;
             case 'metaValues':
-                $metaValues = $this->getCustomPostMetaTypeAPI()->getCustomPostMeta(
+                $metaValues = $this->getLogicalCustomPostMetaTypeAPI()->getCustomPostMeta(
                     $customPost,
                     $fieldDataAccessor->getValue('key'),
                     false
@@ -117,7 +122,7 @@ class CustomPostObjectTypeFieldResolver extends AbstractWithMetaObjectTypeFieldR
                 return $metaValues;
             case 'meta':
                 $meta = [];
-                $allMeta = $this->getCustomPostMetaTypeAPI()->getAllCustomPostMeta($customPost);
+                $allMeta = $this->getLogicalCustomPostMetaTypeAPI()->getAllCustomPostMeta($customPost);
                 /** @var string[] */
                 $keys = $fieldDataAccessor->getValue('keys');
                 foreach ($keys as $key) {