+++ layers/Engine/packages/engine/src/TypeResolvers/ScalarType/AbstractJSONObjectScalarTypeResolver.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoP\Engine\TypeResolvers\ScalarType;
+
+use PoP\ComponentModel\Feedback\ObjectTypeFieldResolutionFeedbackStore;
+use PoP\ComponentModel\TypeResolvers\ScalarType\AbstractScalarTypeResolver;
+use PoP\GraphQLParser\Spec\Parser\Ast\AstInterface;
+use stdClass;
+
+/**
+ * GraphQL Custom Scalar representing a JSON Object on the client-side,
+ * handled via an stdClass object on the server-side
+ *
+ * @see https://spec.graphql.org/draft/#sec-Scalars.Custom-Scalars
+ */
+abstract class AbstractJSONObjectScalarTypeResolver extends AbstractScalarTypeResolver
+{
+    public function coerceValue(
+        string|int|float|bool|stdClass $inputValue,
+        AstInterface $astNode,
+        ObjectTypeFieldResolutionFeedbackStore $objectTypeFieldResolutionFeedbackStore,
+    ): string|int|float|bool|object|null {
+        if (!($inputValue instanceof stdClass)) {
+            $this->addDefaultError($inputValue, $astNode, $objectTypeFieldResolutionFeedbackStore);
+            return null;
+        }
+        return $inputValue;
+    }
+}