+++ layers/Schema/packages/extended-schema-commons/src/TypeResolvers/ScalarType/AbstractScalarListValueJSONObjectScalarTypeResolver.php
@@ -5,11 +5,11 @@ declare(strict_types=1);
 namespace PoPSchema\ExtendedSchemaCommons\TypeResolvers\ScalarType;
 
 use PoP\ComponentModel\Feedback\ObjectTypeFieldResolutionFeedbackStore;
-use PoP\Engine\TypeResolvers\ScalarType\JSONObjectScalarTypeResolver;
+use PoP\Engine\TypeResolvers\ScalarType\AbstractJSONObjectScalarTypeResolver;
 use PoP\GraphQLParser\Spec\Parser\Ast\AstInterface;
 use stdClass;
 
-abstract class AbstractScalarListValueJSONObjectScalarTypeResolver extends JSONObjectScalarTypeResolver
+abstract class AbstractScalarListValueJSONObjectScalarTypeResolver extends AbstractJSONObjectScalarTypeResolver
 {
     public function coerceValue(
         string|int|float|bool|stdClass $inputValue,